+++ R/hardhat.R
@@ -36,7 +36,7 @@
 #'  If no argument is supplied, this will use the default values in [tabnet_config()].
 #' @param from_epoch When a `tabnet_model` is provided, restore the network weights from a specific epoch.
 #'  Default is last available checkpoint for restored model, or last epoch for in-memory model.
-#' @param weights Unused.
+#' @param weights Unused. Placeholder for hardhat::importance_weight() variables.
 #' @param ... Model hyperparameters.
 #' Any hyperparameters set here will update those set by the config argument.
 #' See [tabnet_config()] for a list of all possible hyperparameters.
@@ -68,37 +68,36 @@
 #' torch::torch_set_num_interop_threads(1)
 #' ```
 #'
-#' @examplesIf torch::torch_is_installed()
-#'
+#' @examplesIf (torch::torch_is_installed() && require("modeldata"))
+#' \dontrun{
 #' data("ames", package = "modeldata")
 #' data("attrition", package = "modeldata")
-#' ids <- sample(nrow(attrition), 256)
 #'
 #' ## Single-outcome regression using formula specification
-#' fit <- tabnet_fit(Sale_Price ~ ., data = ames, epochs = 1)
+#' fit <- tabnet_fit(Sale_Price ~ ., data = ames, epochs = 4)
 #'
 #' ## Single-outcome classification using data-frame specification
-#' attrition_x <- attrition[,-which(names(attrition) == "Attrition")]
-#' fit <- tabnet_fit(attrition_x, attrition$Attrition, epochs = 1, verbose = TRUE)
+#' attrition_x <- attrition[ids,-which(names(attrition) == "Attrition")]
+#' fit <- tabnet_fit(attrition_x, attrition$Attrition, epochs = 4, verbose = TRUE)
 #'
 #' ## Multi-outcome regression on `Sale_Price` and `Pool_Area` in `ames` dataset using formula,
-#' ames_fit <- tabnet_fit(Sale_Price + Pool_Area ~ ., data = ames[ids,], epochs = 2, valid_split = 0.2)
+#' ames_fit <- tabnet_fit(Sale_Price + Pool_Area ~ ., data = ames, epochs = 4, valid_split = 0.2)
 #'
 #' ## Multi-label classification on `Attrition` and `JobSatisfaction` in
 #' ## `attrition` dataset using recipe
 #' library(recipes)
-#' rec <- recipe(Attrition + JobSatisfaction ~ ., data = attrition[ids,]) %>%
+#' rec <- recipe(Attrition + JobSatisfaction ~ ., data = attrition) %>%
 #'   step_normalize(all_numeric(), -all_outcomes())
 #'
-#' attrition_fit <- tabnet_fit(rec, data = attrition[ids,], epochs = 2, valid_split = 0.2)
+#' attrition_fit <- tabnet_fit(rec, data = attrition, epochs = 4, valid_split = 0.2)
 #'
 #' ## Hierarchical classification on  `acme`
 #' data(acme, package = "data.tree")
 #'
-#' acme_fit <- tabnet_fit(acme, epochs = 2, verbose = TRUE)
+#' acme_fit <- tabnet_fit(acme, epochs = 4, verbose = TRUE)
 #'
-#' # Note: Dataset number of rows and model number of epochs should be increased
-#' # for publication-level results.
+#' # Note: Model's number of epochs should be increased for publication-level results.
+#' }
 #' @return A TabNet model object. It can be used for serialization, predictions, or further fitting.
 #'
 #' @export
@@ -118,22 +117,10 @@ tabnet_fit.default <- function(x, ...) {
 #' @rdname tabnet_fit
 tabnet_fit.data.frame <- function(x, y, tabnet_model = NULL, config = tabnet_config(), ...,
                                   from_epoch = NULL, weights = NULL) {
-  if (!is.null(weights)) {
-    message(gettextf("Configured `weights` will not be used"))
-  }
   processed <- hardhat::mold(x, y)
   check_type(processed$outcomes)
 
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-    function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-    default_config,
-    new_config)
-    ]
-  config <- utils::modifyList(config, as.list(new_config))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
@@ -141,9 +128,6 @@ tabnet_fit.data.frame <- function(x, y, tabnet_model = NULL, config = tabnet_con
 #' @rdname tabnet_fit
 tabnet_fit.formula <- function(formula, data, tabnet_model = NULL, config = tabnet_config(), ...,
                                from_epoch = NULL, weights = NULL) {
-  if (!is.null(weights)) {
-    message(gettextf("Configured `weights` will not be used"))
-  }
   processed <- hardhat::mold(
     formula, data,
     blueprint = hardhat::default_formula_blueprint(
@@ -153,16 +137,7 @@ tabnet_fit.formula <- function(formula, data, tabnet_model = NULL, config = tabn
   )
   check_type(processed$outcomes)
 
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-      default_config,
-      new_config)
-  ]
-  config <- utils::modifyList(config, as.list(new_config))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
@@ -170,22 +145,10 @@ tabnet_fit.formula <- function(formula, data, tabnet_model = NULL, config = tabn
 #' @rdname tabnet_fit
 tabnet_fit.recipe <- function(x, data, tabnet_model = NULL, config = tabnet_config(), ...,
                               from_epoch = NULL, weights = NULL) {
-  if (!is.null(weights)) {
-    message(gettextf("Configured `weights` will not be used"))
-  }
   processed <- hardhat::mold(x, data)
   check_type(processed$outcomes)
 
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-      default_config,
-      new_config)
-  ]
-  config <- utils::modifyList(config, as.list(new_config))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
@@ -210,16 +173,7 @@ tabnet_fit.Node <- function(x, tabnet_model = NULL, config = tabnet_config(), ..
   ancestor_m <- Matrix::sparseMatrix(ancestor$from, ancestor$to, dims = dims, x = 1)
   check_type(processed$outcomes)
 
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-      default_config,
-      new_config)
-  ]
-  config <- utils::modifyList(config, as.list(new_config, ancestor = ancestor_m))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "supervised")
 }
 
@@ -325,16 +279,7 @@ tabnet_pretrain.default <- function(x, ...) {
 tabnet_pretrain.data.frame <- function(x, y, tabnet_model = NULL, config = tabnet_config(), ..., from_epoch = NULL) {
   processed <- hardhat::mold(x, y)
 
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-      default_config,
-      new_config)
-  ]
-  config <- utils::modifyList(config, as.list(new_config))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "unsupervised")
 }
 
@@ -348,17 +293,7 @@ tabnet_pretrain.formula <- function(formula, data, tabnet_model = NULL, config =
       intercept = FALSE
     )
   )
-
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-      default_config,
-      new_config)
-  ]
-  config <- utils::modifyList(config, as.list(new_config))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "unsupervised")
 }
 
@@ -367,16 +302,7 @@ tabnet_pretrain.formula <- function(formula, data, tabnet_model = NULL, config =
 tabnet_pretrain.recipe <- function(x, data, tabnet_model = NULL, config = tabnet_config(), ..., from_epoch = NULL) {
   processed <- hardhat::mold(x, data)
 
-  default_config <- tabnet_config()
-  new_config <- do.call(tabnet_config, list(...))
-  new_config <- new_config[
-    mapply(
-      function(x, y) ifelse(is.null(x), !is.null(y), x != y),
-      default_config,
-      new_config)
-  ]
-  config <- utils::modifyList(config, as.list(new_config))
-
+  config <- merge_config_and_dots(config, ...)
   tabnet_bridge(processed, config = config, tabnet_model, from_epoch, task = "unsupervised")
 }
 
@@ -548,13 +474,6 @@ predict_tabnet_bridge <- function(type, object, predictors, epoch, batch_size) {
   )
 }
 
-model_to_raw <- function(model) {
-  con <- rawConnection(raw(), open = "wr")
-  torch::torch_save(model, con)
-  on.exit({close(con)}, add = TRUE)
-  r <- rawConnectionValue(con)
-  r
-}
 
 model_pretrain_to_fit <- function(obj, x, y, config = tabnet_config()) {
 
@@ -584,20 +503,6 @@ model_pretrain_to_fit <- function(obj, x, y, config = tabnet_config()) {
   }
   tabnet_model_lst$network$load_state_dict(tabnet_state_dict)
   tabnet_model_lst
-}
-
-
-check_net_is_empty_ptr <- function(object) {
-  is_null_external_pointer(object$fit$network$.check$ptr)
-}
-
-# https://stackoverflow.com/a/27350487/3297472
-is_null_external_pointer <- function(pointer) {
-  a <- attributes(pointer)
-  attributes(pointer) <- NULL
-  out <- identical(pointer, methods::new("externalptr"))
-  attributes(pointer) <- a
-  out
 }
 
 #' Check consistency between modeling-task type and class of outcomes vars.
@@ -643,88 +548,6 @@ check_type <- function(outcome_ptype, type = NULL) {
 }
 
 
-#' Check that Node object names are compliant
-#'
-#' @param node the Node object, or a dataframe ready to be parsed by `data.tree::as.Node()`
-#'
-#' @return node if it is compliant, else an Error with the column names to fix
-#' @export
-#'
-#' @examplesIf (require("data.tree") || require("dplyr"))
-#' library(dplyr)
-#' library(data.tree)
-#' data(starwars)
-#' starwars_tree <- starwars %>%
-#'   mutate(pathString = paste("tree", species, homeworld, `name`, sep = "/"))
-#'
-#' # pre as.Node() check
-#' try(check_compliant_node(starwars_tree))
-#'
-#' # post as.Node() check
-#' check_compliant_node(as.Node(starwars_tree))
-#'
-check_compliant_node <- function(node) {
-  #  prevent reserved data.tree Node colnames and the level_1 ... level_n names used for coercion
-  if (inherits(node, "Node")) {
-    # Node has already lost its reserved colnames
-    reserved_names <- paste0("level_", c(1:node$height))
-    actual_names <- node$attributesAll
-  } else if (inherits(node, "data.frame") && "pathString" %in% colnames(node)) {
-    node_height <- max(stringr::str_count(node$pathString, "/"))
-    reserved_names <- c(paste0("level_", c(1:node_height)), data.tree::NODE_RESERVED_NAMES_CONST)
-    actual_names <- colnames(node)[!colnames(node) %in% "pathString"]
-  } else {
-    stop("The provided hierarchical object is not recognized with a valid format that can be checked", call. = FALSE)
-  }
-
-  if (any(actual_names %in% reserved_names)) {
-    stop(domain=NA,
-         gettextf("The attributes or colnames in the provided hierarchical object use the following reserved names : '%s'. Please change those names as they will lead to unexpected tabnet behavior.",
-          paste(actual_names[actual_names %in% reserved_names], collapse = "', '")
-         ),
-         call. = FALSE)
-  }
-
-  invisible(node)
-}
-
-#' Turn a Node object into predictor and outcome.
-#'
-#' @param x Node object
-#' @param drop_last_level TRUE unused
-#'
-#' @return a named list of x and y, being respectively the predictor data-frame and the outcomes data-frame,
-#'   as expected inputs for `hardhat::mold()` function.
-#' @export
-#'
-#' @examplesIf (require("data.tree") || require("dplyr"))
-#' library(dplyr)
-#' library(data.tree)
-#' data(starwars)
-#' starwars_tree <- starwars %>%
-#'   mutate(pathString = paste("tree", species, homeworld, `name`, sep = "/")) %>%
-#'   as.Node()
-#' node_to_df(starwars_tree)$x %>% head()
-#' node_to_df(starwars_tree)$y %>% head()
-#' @importFrom dplyr last_col mutate mutate_if select starts_with where
-node_to_df <- function(x, drop_last_level = TRUE) {
-  # TODO get rid of all those import through base R equivalent
-  xy_df <- data.tree::ToDataFrameTypeCol(x, x$attributesAll)
-  x_df <- xy_df %>%
-    select(-starts_with("level_")) %>%
-    mutate_if(is.character, as.factor)
-  y_df <- xy_df %>%
-    select(starts_with("level_")) %>%
-    # drop first (and all zero-variance) column
-    select(where(~ nlevels(as.factor(.x)) > 1 )) %>%
-    # TODO take the drop_last_level param into account
-    # drop last level column
-    select(-last_col()) %>%
-    # TODO impute "NA" with parent through coalesce() via an option
-    mutate_if(is.character, as.factor)
-  return(list(x = x_df, y = y_df))
-}
-
 reload_model <- function(object) {
   con <- rawConnection(object)
   on.exit({close(con)}, add = TRUE)
@@ -753,6 +576,16 @@ print.tabnet_pretrain <- print.tabnet_fit
 #'
 #' @return a tabnet network with the top nn_layer removed
 #' @rdname nn_prune_head
+#' @examplesIf (torch::torch_is_installed())
+#' data("ames", package = "modeldata")
+#' x <- ames[,-which(names(ames) == "Sale_Price")]
+#' y <- ames$Sale_Price
+#' # pretrain a tabnet model on ames dataset
+#' ames_pretrain <- tabnet_pretrain(x, y, epoch = 2, checkpoint_epochs = 1)
+#' # prune classification head to get an embedding model
+#' pruned_pretrain <- torch::nn_prune_head(ames_pretrain, 1)
+#
+#' @importFrom torch nn_prune_head
 #' @export
 nn_prune_head.tabnet_fit <- function(x, head_size) {
   if (check_net_is_empty_ptr(x)) {
@@ -761,18 +594,18 @@ nn_prune_head.tabnet_fit <- function(x, head_size) {
     net <- x$fit$network
   }
   # here we assemble nn_prune_head(x, 1) with nn_prune_head(x$tabnet, 1)
-  x <- torch::nn_prune_head(net, 1)
-  x$add_module(name= "tabnet", module=torch::nn_prune_head(net$tabnet,head_size=head_size))
+  x <- nn_prune_head(net, 1)
+  x$add_module(name= "tabnet", module=nn_prune_head(net$tabnet,head_size=head_size))
 
 }
+#' @importFrom torch nn_prune_head
 #' @rdname nn_prune_head
 #' @export
 nn_prune_head.tabnet_pretrain <- function(x, head_size) {
   if (check_net_is_empty_ptr(x)) {
-    torch::nn_prune_head(reload_model(x$serialized_net), head_size=head_size)
+    nn_prune_head(reload_model(x$serialized_net), head_size=head_size)
   } else {
-    torch::nn_prune_head(x$fit$network, head_size=head_size)
+    nn_prune_head(x$fit$network, head_size=head_size)
   }
 
 }
-