+++ R/model.R
@@ -130,6 +130,15 @@ resolve_data <- function(x, y) {
 #' @param skip_importance if feature importance calculation should be skipped (default: `FALSE`)
 #' @return A named list with all hyperparameters of the TabNet implementation.
 #'
+#' @examplesIf (torch::torch_is_installed() && require("modeldata"))
+#' data("ames", package = "modeldata")
+#'
+#' # change the model config for an faster ignite optimizer
+#' config <- tabnet_config(optimizer = torch::optim_ignite_adamw)
+#'
+#' ## Single-outcome regression using formula specification
+#' fit <- tabnet_fit(Sale_Price ~ ., data = ames, epochs = 1, config = config)
+#'
 #' @export
 tabnet_config <- function(batch_size = 1024^2,
                           penalty = 1e-3,
@@ -190,7 +199,7 @@ tabnet_config <- function(batch_size = 1024^2,
     virtual_batch_size = virtual_batch_size,
     valid_split = valid_split,
     learn_rate = learn_rate,
-    optimizer = optimizer,
+    optimizer = resolve_optimizer(optimizer),
     lr_scheduler = lr_scheduler,
     lr_decay = lr_decay,
     step_size = step_size,
@@ -231,7 +240,9 @@ max_constraint_output <- function(output, labels, ancestor) {
 resolve_loss <- function(config, dtype) {
   loss <- config$loss
 
-  if (is.function(loss))
+  if (is_loss_generator(loss))
+    loss_fn <- loss()
+  else if (is.function(loss))
     loss_fn <- loss
   else if (loss %in% c("mse", "auto") && !dtype == torch::torch_long())
     loss_fn <- torch::nn_mse_loss()
@@ -500,18 +511,12 @@ tabnet_train_supervised <- function(obj, x, y, config = tabnet_config(), epoch_s
   if (!is.null(config$ancestor)) {
     config$ancestor_tt <- torch::torch_tensor(config$ancestor)$to(torch::torch_bool(), device = device)
   }
-  # define optimizer
-  if (rlang::is_function(config$optimizer)) {
 
+  # instanciate optimizer
+  if (is_optim_generator(config$optimizer)) {
     optimizer <- config$optimizer(network$parameters, config$learn_rate)
-
-  } else if (rlang::is_scalar_character(config$optimizer)) {
-
-    if (config$optimizer == "adam")
-      optimizer <- torch::optim_adam(network$parameters, lr = config$learn_rate)
-    else
-      stop("Currently only the 'adam' optimizer is supported.", call. = FALSE)
-
+  } else {
+    stop("`optimizer` must be resolved into a torch optimizer generator.", call. = FALSE)
   }
 
   # define scheduler
@@ -648,7 +653,7 @@ predict_impl <- function(obj, x, batch_size = 1e5) {
   device = obj$fit$config$device
   predict_ds <-   torch::dataset(
     initialize = function() {},
-    .getbatch = function(batch) {resolve_data(x[batch,], rep(1, nrow(x)))},
+    .getbatch = function(batch) {resolve_data(x[batch,], matrix(1L, nrow = length(batch)))},
     .length = function() {nrow(x)}
   )()
 
@@ -748,16 +753,4 @@ predict_impl_class_multiple <- function(obj, x, batch_size, outcome_nlevels) {
     ~factor(.y[.x], levels = .y)
   )
   hardhat::spruce_class_multiple(!!!p_factor_lst)
-}
-
-to_device <- function(x, device) {
-  lapply(x, function(x) {
-    if (inherits(x, "torch_tensor")) {
-      x$to(device=device)
-    } else if (is.list(x)) {
-      lapply(x, to_device)
-    } else {
-      x
-    }
-  })
 }